optimizando y haciendo refactoring al codigo
