+++ src/main/java/com/samy/service/app/config/AwsConfig.java
@@ -4,7 +4,6 @@ import org.socialsignin.spring.data.dynamodb.repository.config.EnableDynamoDBRep
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
 
 import com.amazonaws.auth.AWSCredentialsProvider;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
@@ -32,12 +31,6 @@ public class AwsConfig {
 
     @Value("${aws.config.service-endpoint}")
     private String serviceEnpoint;
-
-    
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
     
     @Bean
     public AmazonDynamoDB amazonDynamoDB() {
