+++ src/main/java/com/samy/service/app/service/processor/CasoRequestBuilder.java
@@ -1,4 +1,4 @@
-package com.samy.service.app.service.impl;
+package com.samy.service.app.service.processor;
 
 import static com.samy.service.app.util.ListUtils.listArchivoAdjunto;
 import static com.samy.service.app.util.Utils.uuidGenerado;
